% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{image_stitch}
\alias{image_stitch}
\title{Stitch images}
\usage{
image_stitch(x, config = system.file(package = "image.OpenPano", "extdata",
  "config.cfg"), file = tempfile(fileext = ".jpg"))
}
\arguments{
\item{x}{a character vector of paths to jpeg files to stitch}

\item{config}{the path to the config file of OpenPano}

\item{file}{path to the output file where the stitched image will be written unto. 
This should be a file with extension .jpg}
}
\value{
a list with elements x and file where \code{x} are the input files and \code{file}
is the stitched output file
}
\description{
Stitch images
}
\examples{
folder <- system.file(package = "image.OpenPano", "extdata")
images <- c(file.path(folder, "imga.jpg"), 
            file.path(folder, "imgb.jpg"),
            file.path(folder, "imgc.jpg"))
result <- image_stitch(images, file = "result_stitched.jpg")

library(magick)
image_read(images[1])
image_read(images[2])
image_read(images[3])
image_read("result_stitched.jpg")

file.remove("result_stitched.jpg")
}
